// frontend/types/client-project.ts
export type ClientProject = {
    id: number; 
    title: string;
    subtitle: string; 
    description: string;  
    challenge: string; 
    solution: string; 
    defaultImage: string; 
    hoverImage: string; 
    imageAlt: string; 
    imagePosition: string; 
    badge: string; 
    url: string; 
};